Add DragonFlyBSD support
