+++ process_unix.go
@@ -1,15 +1,13 @@
-// +build linux
+// +build linux dragonfly
 
 package ps
 
 import (
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strconv"
-	"strings"
 )
 
 // UnixProcess is an implementation of Process that contains Unix-specific
@@ -42,38 +40,6 @@ func (p *UnixProcess) Executable() string {
 		return p.binary
 	}
 	return filepath.Base(path)
-}
-
-// Path returns path to process executable
-func (p *UnixProcess) Path() (string, error) {
-	return filepath.EvalSymlinks(fmt.Sprintf("/proc/%d/exe", p.pid))
-}
-
-// Refresh reloads all the data associated with this process.
-func (p *UnixProcess) Refresh() error {
-	statPath := fmt.Sprintf("/proc/%d/stat", p.pid)
-	dataBytes, err := ioutil.ReadFile(statPath)
-	if err != nil {
-		return err
-	}
-
-	// First, parse out the image name
-	data := string(dataBytes)
-	binStart := strings.IndexRune(data, '(') + 1
-	binEnd := strings.IndexRune(data[binStart:], ')')
-	p.binary = data[binStart : binStart+binEnd]
-
-	// Move past the image name and start parsing the rest
-	// The name here might not be the full name
-	data = data[binStart+binEnd+2:]
-	_, err = fmt.Sscanf(data,
-		"%c %d %d %d",
-		&p.state,
-		&p.ppid,
-		&p.pgrp,
-		&p.sid)
-
-	return err
 }
 
 func findProcess(pid int) (Process, error) {
