+++ process_unix_test.go
@@ -1,4 +1,4 @@
-// +build linux
+// +build linux dragonfly
 
 package ps
